Address additional PR feedback
